02/15
 ** Player **
1: 增加 Health
2: 新增 Function
 - TakeDamage
 - Die
 - UpdateHpBar
3: Breath 增加 isAim > Breath(0,0)

 ** RPC **
1: Weapon > TakeDamage

 ** Map **
1: 增加障礙物
2: 增加重生點
3: 幫角色和障礙物上物理材質

 ** Canvas **
1: 新增 image > HealthBar

 ** Question **
1: 不知道為甚麼兩台電腦生成的彈孔不同步

+++ FPS_Version2/Assets/1.1_Scripts/Weapon/scr_WeaponData.cs
@@ -3,12 +3,11 @@
 [CreateAssetMenu(fileName = "New gun", menuName = "Gun")]
 public class scr_WeaponData : ScriptableObject
 {
-
     [Header("武器名稱")] public string name;
 
     [Header("瞄準速度")] public float aimSpeed;
     [Header("開槍頻率")] public float fireRate;
-    [Header("武器傷害")] public float damage;
+    [Header("武器傷害")] public int damage;
 
     [Header("武器預置物")] public GameObject weaponPrefab;
 }
